Refactor plugin engine
* Extract websocket connection package. This package holds the server
  and websocket event handlers. A connection class is used to gather
  the created connection.
* Extract setup module, for setting up the engine.
* Extract a const module with constants.
* Extract helper module with helper functions that are aware of the
  engine.
* Extract plugin handler module.
* Extract more util functions to the util package.
